+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout-collection.component.spec.js
@@ -1,38 +1,35 @@
 import forEach from 'mocha-each';
-import DemoControls from '../../../test-utils/component-objects/demo-controls';
 import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
 
 describe('f-checkout "collection" component tests', () => {
-      beforeEach(() => {
+    before(() => {
         browser.url('?path=/story/components-organisms--checkout-component');
+        CheckoutComponent.changeCheckoutMethod('collection');
         browser.switchToFrame(0);
         CheckoutComponent.waitForCheckoutComponent();
     });
 
-    it.skip('should display "mobileNumber" error message when collection method is set', () => {
+    it('should display "mobileNumber" error message when collection method is set and number is incorrect', () => {
         // Arrange
-        DemoControls.selectCheckoutMethod('collection');
+        const addressDetails = {
+            mobileNumber: '1234'
+        };
 
         // Act
-        CheckoutComponent.submit();
+        CheckoutComponent.populateCollectionCheckoutForm(addressDetails);
+        CheckoutComponent.goToPayment();
 
         // Assert
         expect(CheckoutComponent.isFieldErrorDisplayed('mobileNumber')).toBe(true);
     });
 
     forEach(['addressLine1', 'addressLine2', 'addressCity', 'addressPostcode'])
-    .it.skip('address fields should not exist', field => {
-        // Arrange
-        DemoControls.selectCheckoutMethod('collection');
-
+    .it('address fields should not exist', field => {
         // Assert
-        expect(CheckoutComponent.doesInputFieldExist(field)).toBe(false);
+        expect(CheckoutComponent.doesFieldExist(field)).toBe(false);
     });
 
-    it.skip('should display the mandatory "mobileNumber" field', () => {
-        // Arrange
-        DemoControls.selectCheckoutMethod('collection');
-
+    it('should display the mandatory "mobileNumber" field', () => {
         // Assert
         expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
     });