+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,17 +3,32 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v0.31.0
+v0.32.0
 ------------------------------
 *January 5, 2021*
 
 ### Changed
 - Tests for `Checkout` component methods and computed values.
-- Updated config for latest `sass-loader`.
-- Updated fozzie dependencies in `common.scss` to pull in v5-beta.
 
 ### Added
 - `it-IT` tenant file.
+
+
+v0.31.0
+-------------------------------
+*January 5, 2021*
+
+### Changed
+- Updated skipped tests to work with Storybook component tests.
+
+### Added
+- New `clearCheckoutForm` function added to clear the fields in tests in the `f-checkout-component.js` file.
+
+*December 30, 2020*
+
+### Changed
+- Updated config for latest `sass-loader`.
+- Updated fozzie dependencies in `common.scss` to pull in v5-beta.
 
 
 v0.30.0